Revert cbna_inference changes - this branch is only for pooling3d
